重新定义了bwic和bond的表的结构：Bwic加入了上次bid的交易日期以及当前的最高交易金额。而为了方便进行用户推荐，bond加入了transaiction_counts属性，并且在createBid的时候进行更新

+++ service/bwic-api/src/main/java/com/ebidding/bwic/api/BwicDTO.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @Builder
 public class BwicDTO {
     private String bondId;
-    private BigDecimal startPrice;
+    private Double startPrice;
     private LocalDateTime startTime;
     private LocalDateTime dueTime;
     private double size;
